chore: rename server to 'curfu'
