+++ server/curfu/__init__.py
@@ -12,9 +12,9 @@ APP_ROOT = Path(__file__).resolve().parents[0]
 # establish environment-dependent params
 if "FUSION_EB_PROD" in environ:
     environ["FUSION_EB_PROD"] = "true"
-    LOG_FN = "/tmp/fusion_backend.log"
+    LOG_FN = "/tmp/curfu.log"
 else:
-    LOG_FN = "fusion_backend.log"
+    LOG_FN = "curfu.log"
 
 # set up logging
 logging.basicConfig(
