+++ server/curfu/gene_services.py
@@ -6,7 +6,7 @@ from gene.query import QueryHandler
 from gene.schemas import MatchType
 from ga4gh.vrsatile.pydantic.vrsatile_models import CURIE
 
-from curation import logger, ServiceWarning, MAX_SUGGESTIONS, APP_ROOT
+from curfu import logger, ServiceWarning, MAX_SUGGESTIONS, APP_ROOT
 
 
 # term -> (normalized ID, normalized label)
