+++ server/curfu/sequence_services.py
@@ -1,7 +1,7 @@
 """Provide sequence ID generation services."""
 import logging
 
-logger = logging.getLogger("curation_backend")
+logger = logging.getLogger("curfu")
 logger.setLevel(logging.DEBUG)
