+++ server/curfu/version.py
