+++ server/tests/integration/test_main.py
@@ -16,7 +16,7 @@ async def test_service_info(async_client):
     response_json = response.json()
     assert response_json["warnings"] == []
     SEMVER_PATTERN = r"^(?P<major>0|[1-9]\d*)\.(?P<minor>0|[1-9]\d*)\.(?P<patch>0|[1-9]\d*)(?:-(?P<prerelease>(?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\.(?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\+(?P<buildmetadata>[0-9a-zA-Z-]+(?:\.[0-9a-zA-Z-]+)*))?$"  # noqa: E501
-    assert re.match(SEMVER_PATTERN, response_json["fusion_curation_version"])
+    assert re.match(SEMVER_PATTERN, response_json["curfu_version"])
     assert re.match(SEMVER_PATTERN, response_json["fusor_version"])
     assert re.match(SEMVER_PATTERN, response_json["uta_tools_version"])
     # not sure if I want to include vrs-python
